feat: 🎸 調整其他活動查看更多功能、拆分 event card 避免其他活動取得資料時重複渲染其他元件
調整最新、熱門活動 SWR 設定避免更新其他活動時觸發重複取得熱門、最新資料
